refactor: split project service helpers
