+++ pkg/agentcompose/project_service.go
@@ -6,13 +6,11 @@ import (
 	"encoding/json"
 	"errors"
 	"fmt"
-	"path/filepath"
 	"slices"
 	"strings"
 	"time"
 
 	"connectrpc.com/connect"
-	"gopkg.in/yaml.v3"
 
 	"agent-compose/pkg/compose"
 	driverpkg "agent-compose/pkg/driver"
@@ -358,290 +356,6 @@ func (s *Service) resolveProjectRef(ctx context.Context, ref *agentcomposev2.Pro
 		return ProjectRecord{}, fmt.Errorf("project name %s is ambiguous; use project_id or source_path", name)
 	}
 	return matches[0], nil
-}
-
-func normalizeProjectServiceSpec(spec *agentcomposev2.ProjectSpec, source *agentcomposev2.ProjectSource, expectedHash string) (normalizedV2Project, []*agentcomposev2.ProjectValidationIssue, error) {
-	if spec == nil {
-		return normalizedV2Project{}, []*agentcomposev2.ProjectValidationIssue{projectValidationIssue("spec", "project spec is required")}, nil
-	}
-	raw, issues := projectSpecYAMLShape(spec)
-	if len(issues) > 0 {
-		return normalizedV2Project{}, issues, nil
-	}
-	data, err := yaml.Marshal(raw)
-	if err != nil {
-		return normalizedV2Project{}, nil, fmt.Errorf("marshal project spec: %w", err)
-	}
-	parsed, err := compose.Parse(data)
-	if err != nil {
-		return normalizedV2Project{}, []*agentcomposev2.ProjectValidationIssue{issueFromComposeError(err)}, nil
-	}
-	sourcePath := projectServiceSourcePath(source)
-	normalized, err := compose.Normalize(parsed, compose.NormalizeOptions{
-		ComposePath: sourcePath,
-		ProjectDir:  strings.TrimSpace(source.GetProjectDir()),
-	})
-	if err != nil {
-		return normalizedV2Project{}, []*agentcomposev2.ProjectValidationIssue{issueFromComposeError(err)}, nil
-	}
-	hash, err := normalized.Hash()
-	if err != nil {
-		return normalizedV2Project{}, nil, fmt.Errorf("hash project spec: %w", err)
-	}
-	result := normalizedV2Project{
-		spec:       normalized,
-		specProto:  ProjectSpecResponse(normalized),
-		specHash:   hash,
-		sourcePath: sourcePath,
-	}
-	expectedHash = strings.TrimSpace(expectedHash)
-	if expectedHash != "" && expectedHash != hash {
-		return result, []*agentcomposev2.ProjectValidationIssue{projectValidationIssue("expected_spec_hash", fmt.Sprintf("expected spec hash %s does not match normalized spec hash %s", expectedHash, hash))}, nil
-	}
-	return result, nil, nil
-}
-
-func projectSpecYAMLShape(spec *agentcomposev2.ProjectSpec) (map[string]any, []*agentcomposev2.ProjectValidationIssue) {
-	root := map[string]any{}
-	if strings.TrimSpace(spec.GetName()) != "" {
-		root["name"] = spec.GetName()
-	}
-	if variables, issues := envVarYAMLMap("variables", spec.GetVariables()); len(issues) > 0 {
-		return nil, issues
-	} else if len(variables) > 0 {
-		root["variables"] = variables
-	}
-	if workspace := workspaceYAMLShape(spec.GetWorkspace()); len(workspace) > 0 {
-		root["workspace"] = workspace
-	}
-	if agents, issues := agentYAMLMap(spec.GetAgents()); len(issues) > 0 {
-		return nil, issues
-	} else if len(agents) > 0 {
-		root["agents"] = agents
-	}
-	if network := networkYAMLShape(spec.GetNetwork()); len(network) > 0 {
-		root["network"] = network
-	}
-	return root, nil
-}
-
-func envVarYAMLMap(path string, vars []*agentcomposev2.EnvVarSpec) (map[string]any, []*agentcomposev2.ProjectValidationIssue) {
-	values := make(map[string]any, len(vars))
-	for i, env := range vars {
-		name := strings.TrimSpace(env.GetName())
-		if _, ok := values[name]; ok {
-			return nil, []*agentcomposev2.ProjectValidationIssue{projectValidationIssue(fmt.Sprintf("%s[%d].name", path, i), fmt.Sprintf("duplicate environment variable %q", name))}
-		}
-		if env.GetSecret() {
-			values[name] = map[string]any{
-				"value":  env.GetValue(),
-				"secret": true,
-			}
-		} else {
-			values[name] = env.GetValue()
-		}
-	}
-	return values, nil
-}
-
-func agentYAMLMap(agents []*agentcomposev2.AgentSpec) (map[string]any, []*agentcomposev2.ProjectValidationIssue) {
-	values := make(map[string]any, len(agents))
-	for i, agent := range agents {
-		name := strings.TrimSpace(agent.GetName())
-		if _, ok := values[name]; ok {
-			return nil, []*agentcomposev2.ProjectValidationIssue{projectValidationIssue(fmt.Sprintf("agents[%d].name", i), fmt.Sprintf("duplicate agent %q", name))}
-		}
-		raw := map[string]any{}
-		if strings.TrimSpace(agent.GetProvider()) != "" {
-			raw["provider"] = agent.GetProvider()
-		}
-		if strings.TrimSpace(agent.GetModel()) != "" {
-			raw["model"] = agent.GetModel()
-		}
-		if agent.GetSystemPrompt() != "" {
-			raw["system_prompt"] = agent.GetSystemPrompt()
-		}
-		if strings.TrimSpace(agent.GetImage()) != "" {
-			raw["image"] = agent.GetImage()
-		}
-		if driver, issues := driverYAMLShape(fmt.Sprintf("agents[%d].driver", i), agent.GetDriver()); len(issues) > 0 {
-			return nil, issues
-		} else if len(driver) > 0 {
-			raw["driver"] = driver
-		}
-		if env, issues := envVarYAMLMap(fmt.Sprintf("agents[%d].env", i), agent.GetEnv()); len(issues) > 0 {
-			return nil, issues
-		} else if len(env) > 0 {
-			raw["env"] = env
-		}
-		if capsetIDs := normalizeCapsetIDs(agent.GetCapsetIds()); len(capsetIDs) > 0 {
-			raw["capset_ids"] = capsetIDs
-		}
-		if workspace := workspaceYAMLShape(agent.GetWorkspace()); len(workspace) > 0 {
-			raw["workspace"] = workspace
-		}
-		if scheduler := schedulerYAMLShape(agent.GetScheduler()); len(scheduler) > 0 {
-			raw["scheduler"] = scheduler
-		}
-		values[name] = raw
-	}
-	return values, nil
-}
-
-func driverYAMLShape(path string, driver *agentcomposev2.DriverSpec) (map[string]any, []*agentcomposev2.ProjectValidationIssue) {
-	if driver == nil {
-		return nil, nil
-	}
-	byName := strings.ToLower(strings.TrimSpace(driver.GetName()))
-	runtimes := make(map[string]any, 3)
-	if driver.GetBoxlite() != nil {
-		runtimes[compose.DriverBoxlite] = map[string]any{
-			"kernel": driver.GetBoxlite().GetKernel(),
-			"rootfs": driver.GetBoxlite().GetRootfs(),
-		}
-	}
-	if driver.GetDocker() != nil {
-		runtimes[compose.DriverDocker] = map[string]any{"host": driver.GetDocker().GetHost()}
-	}
-	if driver.GetMicrosandbox() != nil {
-		runtimes[compose.DriverMicrosandbox] = map[string]any{"profile": driver.GetMicrosandbox().GetProfile()}
-	}
-	switch byName {
-	case "":
-	case compose.DriverBoxlite, compose.DriverDocker, compose.DriverMicrosandbox:
-		for runtimeName := range runtimes {
-			if runtimeName != byName {
-				return nil, []*agentcomposev2.ProjectValidationIssue{projectValidationIssue(path, fmt.Sprintf("driver name %q conflicts with %q runtime config", byName, runtimeName))}
-			}
-		}
-		if existing, ok := runtimes[byName]; ok {
-			return map[string]any{byName: existing}, nil
-		}
-		return map[string]any{byName: map[string]any{}}, nil
-	default:
-		return nil, []*agentcomposev2.ProjectValidationIssue{projectValidationIssue(path+".name", fmt.Sprintf("unsupported runtime driver %q", byName))}
-	}
-	return runtimes, nil
-}
-
-func schedulerYAMLShape(scheduler *agentcomposev2.SchedulerSpec) map[string]any {
-	if scheduler == nil {
-		return nil
-	}
-	raw := map[string]any{"enabled": scheduler.GetEnabled()}
-	triggers := make([]map[string]any, 0, len(scheduler.GetTriggers()))
-	for _, trigger := range scheduler.GetTriggers() {
-		triggers = append(triggers, triggerYAMLShape(trigger))
-	}
-	if len(triggers) > 0 {
-		raw["triggers"] = triggers
-	}
-	if scheduler.GetScript() != "" {
-		raw["script"] = scheduler.GetScript()
-	}
-	return raw
-}
-
-func triggerYAMLShape(trigger *agentcomposev2.TriggerSpec) map[string]any {
-	raw := map[string]any{}
-	if strings.TrimSpace(trigger.GetName()) != "" {
-		raw["name"] = trigger.GetName()
-	}
-	if trigger.GetPrompt() != "" {
-		raw["prompt"] = trigger.GetPrompt()
-	}
-	kind := strings.ToLower(strings.TrimSpace(trigger.GetKind()))
-	if kind == "" || kind == "cron" {
-		if kind == "cron" || strings.TrimSpace(trigger.GetCron()) != "" {
-			raw["cron"] = trigger.GetCron()
-		}
-	}
-	if kind == "" || kind == "interval" {
-		if kind == "interval" || strings.TrimSpace(trigger.GetInterval()) != "" {
-			raw["interval"] = trigger.GetInterval()
-		}
-	}
-	if kind == "" || kind == "timeout" {
-		if kind == "timeout" || strings.TrimSpace(trigger.GetTimeout()) != "" {
-			raw["timeout"] = trigger.GetTimeout()
-		}
-	}
-	if kind == "" || kind == "event" {
-		if kind == "event" || trigger.GetEvent() != nil {
-			raw["event"] = map[string]any{"topic": trigger.GetEvent().GetTopic()}
-		}
-	}
-	if kind != "" && kind != "cron" && kind != "interval" && kind != "timeout" && kind != "event" {
-		raw[kind] = ""
-	}
-	return raw
-}
-
-func workspaceYAMLShape(workspace *agentcomposev2.WorkspaceSpec) map[string]any {
-	if workspace == nil {
-		return nil
-	}
-	raw := map[string]any{}
-	if strings.TrimSpace(workspace.GetProvider()) != "" {
-		raw["provider"] = workspace.GetProvider()
-	}
-	if strings.TrimSpace(workspace.GetUrl()) != "" {
-		raw["url"] = workspace.GetUrl()
-	}
-	if strings.TrimSpace(workspace.GetBranch()) != "" {
-		raw["branch"] = workspace.GetBranch()
-	}
-	if strings.TrimSpace(workspace.GetPath()) != "" {
-		raw["path"] = workspace.GetPath()
-	}
-	return raw
-}
-
-func networkYAMLShape(network *agentcomposev2.NetworkSpec) map[string]any {
-	if network == nil {
-		return nil
-	}
-	return map[string]any{"mode": network.GetMode()}
-}
-
-func projectServiceSourcePath(source *agentcomposev2.ProjectSource) string {
-	if source == nil {
-		return ""
-	}
-	if composePath := strings.TrimSpace(source.GetComposePath()); composePath != "" {
-		return composePath
-	}
-	if projectDir := strings.TrimSpace(source.GetProjectDir()); projectDir != "" {
-		return filepath.Join(projectDir, "agent-compose.yml")
-	}
-	return ""
-}
-
-func issueFromComposeError(err error) *agentcomposev2.ProjectValidationIssue {
-	var validationErr *compose.ValidationError
-	if errors.As(err, &validationErr) {
-		return projectValidationIssue(validationErr.Path, validationErr.Message)
-	}
-	var parseErr *compose.ParseError
-	if errors.As(err, &parseErr) {
-		return projectValidationIssue(parseErr.Path, parseErr.Message)
-	}
-	return projectValidationIssue("spec", err.Error())
-}
-
-func projectValidationIssue(path, message string) *agentcomposev2.ProjectValidationIssue {
-	if strings.TrimSpace(path) == "" {
-		path = "spec"
-	}
-	return &agentcomposev2.ProjectValidationIssue{
-		Severity: agentcomposev2.ProjectValidationSeverity_PROJECT_VALIDATION_SEVERITY_ERROR,
-		Path:     path,
-		Message:  message,
-	}
-}
-
-func specHashOrEmpty(normalized normalizedV2Project) string {
-	return normalized.specHash
 }
 
 func projectAgentRecordsFromSpec(projectID string, revision int64, spec *compose.NormalizedProjectSpec) ([]ProjectAgentRecord, error) {
@@ -1501,206 +1215,4 @@ func agentChangeAction(existing ProjectAgentRecord, found bool, current ProjectA
 		return agentcomposev2.ProjectChangeAction_PROJECT_CHANGE_ACTION_UNCHANGED
 	}
 	return agentcomposev2.ProjectChangeAction_PROJECT_CHANGE_ACTION_UPDATED
-}
-
-func projectResponse(project ProjectRecord, spec *agentcomposev2.ProjectSpec, agents []ProjectAgentRecord, schedulers []ProjectSchedulerRecord) *agentcomposev2.Project {
-	return &agentcomposev2.Project{
-		Summary:    projectSummaryResponse(project, agents, schedulers),
-		Spec:       spec,
-		Agents:     projectAgentResponses(agents),
-		Schedulers: projectSchedulerResponses(schedulers),
-	}
-}
-
-func projectSummaryResponse(project ProjectRecord, agents []ProjectAgentRecord, schedulers []ProjectSchedulerRecord) *agentcomposev2.ProjectSummary {
-	return &agentcomposev2.ProjectSummary{
-		ProjectId:       project.ID,
-		Name:            project.Name,
-		SourcePath:      project.SourcePath,
-		CurrentRevision: uint64(project.CurrentRevision),
-		SpecHash:        project.SpecHash,
-		AgentCount:      uint32(len(agents)),
-		SchedulerCount:  uint32(len(schedulers)),
-		CreatedAt:       formatProjectTime(project.CreatedAt),
-		UpdatedAt:       formatProjectTime(project.UpdatedAt),
-		RemovedAt:       formatProjectTime(project.RemovedAt),
-	}
-}
-
-func projectRevisionResponse(revision ProjectRevisionRecord, spec *agentcomposev2.ProjectSpec) *agentcomposev2.ProjectRevision {
-	return &agentcomposev2.ProjectRevision{
-		ProjectId: revision.ProjectID,
-		Revision:  uint64(revision.Revision),
-		SpecHash:  revision.SpecHash,
-		Spec:      spec,
-		CreatedAt: formatProjectTime(revision.CreatedAt),
-	}
-}
-
-func projectAgentResponses(agents []ProjectAgentRecord) []*agentcomposev2.ProjectAgent {
-	items := make([]*agentcomposev2.ProjectAgent, 0, len(agents))
-	for _, agent := range agents {
-		items = append(items, &agentcomposev2.ProjectAgent{
-			ProjectId:        agent.ProjectID,
-			AgentName:        agent.AgentName,
-			ManagedAgentId:   agent.ManagedAgentID,
-			Provider:         agent.Provider,
-			Model:            agent.Model,
-			Image:            agent.Image,
-			Driver:           agent.Driver,
-			SchedulerEnabled: agent.SchedulerEnabled,
-		})
-	}
-	return items
-}
-
-func projectSchedulerResponses(schedulers []ProjectSchedulerRecord) []*agentcomposev2.ProjectScheduler {
-	items := make([]*agentcomposev2.ProjectScheduler, 0, len(schedulers))
-	for _, scheduler := range schedulers {
-		items = append(items, &agentcomposev2.ProjectScheduler{
-			ProjectId:       scheduler.ProjectID,
-			AgentName:       scheduler.AgentName,
-			SchedulerId:     scheduler.SchedulerID,
-			ManagedLoaderId: scheduler.ManagedLoaderID,
-			Enabled:         scheduler.Enabled,
-			TriggerCount:    uint32(scheduler.TriggerCount),
-		})
-	}
-	return items
-}
-
-// ProjectSpecResponse converts a normalized compose spec into the v2 ProjectSpec API shape.
-func ProjectSpecResponse(spec *compose.NormalizedProjectSpec) *agentcomposev2.ProjectSpec {
-	if spec == nil {
-		return nil
-	}
-	return &agentcomposev2.ProjectSpec{
-		Name:      spec.Name,
-		Variables: envVarResponses(spec.Variables),
-		Workspace: workspaceResponse(spec.Workspace),
-		Agents:    agentSpecResponses(spec.Agents),
-		Network:   networkResponse(spec.Network),
-	}
-}
-
-func agentSpecResponses(agents []compose.NormalizedAgentSpec) []*agentcomposev2.AgentSpec {
-	items := make([]*agentcomposev2.AgentSpec, 0, len(agents))
-	for _, agent := range agents {
-		items = append(items, &agentcomposev2.AgentSpec{
-			Name:         agent.Name,
-			Provider:     agent.Provider,
-			Model:        agent.Model,
-			SystemPrompt: agent.SystemPrompt,
-			Image:        agent.Image,
-			Driver:       driverResponse(agent.Driver),
-			Env:          envVarResponses(agent.Env),
-			CapsetIds:    normalizeCapsetIDs(agent.CapsetIDs),
-			Workspace:    workspaceResponse(agent.Workspace),
-			Scheduler:    schedulerResponse(agent.Scheduler),
-		})
-	}
-	return items
-}
-
-func envVarResponses(values map[string]compose.EnvVarSpec) []*agentcomposev2.EnvVarSpec {
-	names := make([]string, 0, len(values))
-	for name := range values {
-		names = append(names, name)
-	}
-	slices.Sort(names)
-	items := make([]*agentcomposev2.EnvVarSpec, 0, len(values))
-	for _, name := range names {
-		value := values[name]
-		items = append(items, &agentcomposev2.EnvVarSpec{Name: name, Value: value.Value, Secret: value.Secret})
-	}
-	return items
-}
-
-func workspaceResponse(workspace *compose.WorkspaceSpec) *agentcomposev2.WorkspaceSpec {
-	if workspace == nil {
-		return nil
-	}
-	return &agentcomposev2.WorkspaceSpec{
-		Provider: workspace.Provider,
-		Url:      workspace.URL,
-		Branch:   workspace.Branch,
-		Path:     workspace.Path,
-	}
-}
-
-func networkResponse(network *compose.NetworkSpec) *agentcomposev2.NetworkSpec {
-	if network == nil {
-		return nil
-	}
-	return &agentcomposev2.NetworkSpec{Mode: network.Mode}
-}
-
-func driverResponse(driver *compose.NormalizedDriverSpec) *agentcomposev2.DriverSpec {
-	if driver == nil {
-		return nil
-	}
-	result := &agentcomposev2.DriverSpec{Name: driver.Name}
-	switch driver.Name {
-	case compose.DriverBoxlite:
-		result.Boxlite = &agentcomposev2.BoxliteDriverSpec{}
-		if driver.Boxlite != nil {
-			result.Boxlite.Kernel = driver.Boxlite.Kernel
-			result.Boxlite.Rootfs = driver.Boxlite.Rootfs
-		}
-	case compose.DriverDocker:
-		result.Docker = &agentcomposev2.DockerDriverSpec{}
-		if driver.Docker != nil {
-			result.Docker.Host = driver.Docker.Host
-		}
-	case compose.DriverMicrosandbox:
-		result.Microsandbox = &agentcomposev2.MicrosandboxDriverSpec{}
-		if driver.Microsandbox != nil {
-			result.Microsandbox.Profile = driver.Microsandbox.Profile
-		}
-	}
-	return result
-}
-
-func schedulerResponse(scheduler *compose.NormalizedSchedulerSpec) *agentcomposev2.SchedulerSpec {
-	if scheduler == nil {
-		return nil
-	}
-	triggers := make([]*agentcomposev2.TriggerSpec, 0, len(scheduler.Triggers))
-	for _, trigger := range scheduler.Triggers {
-		triggers = append(triggers, triggerResponse(trigger))
-	}
-	return &agentcomposev2.SchedulerSpec{
-		Enabled:  scheduler.Enabled,
-		Triggers: triggers,
-		Script:   scheduler.Script,
-	}
-}
-
-func triggerResponse(trigger compose.NormalizedTriggerSpec) *agentcomposev2.TriggerSpec {
-	result := &agentcomposev2.TriggerSpec{
-		Name:   trigger.Name,
-		Kind:   trigger.Kind,
-		Prompt: trigger.Prompt,
-	}
-	switch trigger.Kind {
-	case "cron":
-		result.Cron = trigger.Cron
-	case "interval":
-		result.Interval = trigger.Interval
-	case "timeout":
-		result.Timeout = trigger.Timeout
-	case "event":
-		result.Event = &agentcomposev2.EventTriggerSpec{}
-		if trigger.Event != nil {
-			result.Event.Topic = trigger.Event.Topic
-		}
-	}
-	return result
-}
-
-func formatProjectTime(value time.Time) string {
-	if value.IsZero() {
-		return ""
-	}
-	return value.UTC().Format(time.RFC3339Nano)
 }
